🎨 Better SQL logging in tests.
