+++ scripts/import_from_application.py
@@ -1,0 +1,9 @@
+#!/usr/bin/env python3
+from app import app
+
+with app.app_context():
+
+    # Grab data from app store.
+
+    # Upsert data into assessment store.
+
